Revert "Rename all test files to use first version instead of last version"
This reverts commit ada63cd5ca5105cd0845e13acd59ee5297b88483.
