+++ functionaltests/8_17_test.go
@@ -19,24 +19,60 @@ package functionaltests
 
 import (
 	"testing"
+
+	"github.com/elastic/apm-server/functionaltests/internal/asserts"
 )
 
-func TestUpgrade_8_17_to_8_18_Snapshot(t *testing.T) {
+// In 8.15, the data stream management was migrated from ILM to DSL.
+// However, a bug was introduced, causing data streams to be unmanaged.
+// See https://github.com/elastic/apm-server/issues/13898.
+//
+// It was fixed by defaulting data stream management to DSL, and eventually
+// reverted back to ILM in 8.17. Therefore, data streams created in 8.15 and
+// 8.16 are managed by DSL instead of ILM.
+
+func TestUpgrade_8_16_to_8_17_Snapshot(t *testing.T) {
 	t.Parallel()
-	from := getLatestSnapshot(t, "8.17")
-	to := getLatestSnapshot(t, "8.18")
+	from := getLatestSnapshot(t, "8.16")
+	to := getLatestSnapshot(t, "8.17")
 	if !from.CanUpgradeTo(to.Version) {
 		t.Skipf("upgrade from %s to %s is not allowed", from.Version, to.Version)
 		return
 	}
 
-	scenarios := basicUpgradeILMTestScenarios(
+	scenarios := allBasicUpgradeScenarios(
 		from.Version,
 		to.Version,
+		// Data streams managed by DSL pre-upgrade.
+		asserts.CheckDataStreamsWant{
+			Quantity:         8,
+			PreferIlm:        false,
+			DSManagedBy:      managedByDSL,
+			IndicesPerDS:     1,
+			IndicesManagedBy: []string{managedByDSL},
+		},
+		// Data streams managed by ILM post-upgrade.
+		// However, the index created before upgrade is still managed by DSL.
+		asserts.CheckDataStreamsWant{
+			Quantity:         8,
+			PreferIlm:        true,
+			DSManagedBy:      managedByILM,
+			IndicesPerDS:     1,
+			IndicesManagedBy: []string{managedByDSL},
+		},
+		// Verify lazy rollover happened, i.e. 2 indices.
+		asserts.CheckDataStreamsWant{
+			Quantity:         8,
+			PreferIlm:        true,
+			DSManagedBy:      managedByILM,
+			IndicesPerDS:     2,
+			IndicesManagedBy: []string{managedByDSL, managedByILM},
+		},
 		apmErrorLogs{
 			tlsHandshakeError,
 			esReturnedUnknown503,
 			refreshCache503,
+			refreshCacheCtxCanceled,
 			populateSourcemapFetcher403,
 		},
 	)
@@ -48,22 +84,48 @@ func TestUpgrade_8_17_to_8_18_Snapshot(t *testing.T) {
 	}
 }
 
-func TestUpgrade_8_17_to_8_18_BC(t *testing.T) {
+func TestUpgrade_8_16_to_8_17_BC(t *testing.T) {
 	t.Parallel()
-	from := getLatestVersionOrSkip(t, "8.17")
-	to := getLatestBCOrSkip(t, "8.18")
+	from := getLatestVersionOrSkip(t, "8.16")
+	to := getLatestBCOrSkip(t, "8.17")
 	if !from.CanUpgradeTo(to.Version) {
 		t.Skipf("upgrade from %s to %s is not allowed", from.Version, to.Version)
 		return
 	}
 
-	scenarios := basicUpgradeILMTestScenarios(
+	scenarios := allBasicUpgradeScenarios(
 		from.Version,
 		to.Version,
+		// Data streams managed by DSL pre-upgrade.
+		asserts.CheckDataStreamsWant{
+			Quantity:         8,
+			PreferIlm:        false,
+			DSManagedBy:      managedByDSL,
+			IndicesPerDS:     1,
+			IndicesManagedBy: []string{managedByDSL},
+		},
+		// Data streams managed by ILM post-upgrade.
+		// However, the index created before upgrade is still managed by DSL.
+		asserts.CheckDataStreamsWant{
+			Quantity:         8,
+			PreferIlm:        true,
+			DSManagedBy:      managedByILM,
+			IndicesPerDS:     1,
+			IndicesManagedBy: []string{managedByDSL},
+		},
+		// Verify lazy rollover happened, i.e. 2 indices.
+		asserts.CheckDataStreamsWant{
+			Quantity:         8,
+			PreferIlm:        true,
+			DSManagedBy:      managedByILM,
+			IndicesPerDS:     2,
+			IndicesManagedBy: []string{managedByDSL, managedByILM},
+		},
 		apmErrorLogs{
 			tlsHandshakeError,
 			esReturnedUnknown503,
 			refreshCache503,
+			refreshCacheCtxCanceled,
 			populateSourcemapFetcher403,
 		},
 	)
