+++ functionaltests/8_18_test.go
@@ -21,21 +21,16 @@ import (
 	"testing"
 )
 
-// Data streams get marked for lazy rollover by ES when something
-// changed in the underlying template(s), which in this case is
-// the apm-data plugin update for 8.19 and 9.1:
-// https://github.com/elastic/elasticsearch/pull/119995.
-
-func TestUpgrade_8_18_to_8_19_Snapshot(t *testing.T) {
+func TestUpgrade_8_17_to_8_18_Snapshot(t *testing.T) {
 	t.Parallel()
-	from := getLatestSnapshot(t, "8.18")
-	to := getLatestSnapshot(t, "8.19")
+	from := getLatestSnapshot(t, "8.17")
+	to := getLatestSnapshot(t, "8.18")
 	if !from.CanUpgradeTo(to.Version) {
 		t.Skipf("upgrade from %s to %s is not allowed", from.Version, to.Version)
 		return
 	}
 
-	scenarios := basicUpgradeLazyRolloverILMTestScenarios(
+	scenarios := basicUpgradeILMTestScenarios(
 		from.Version,
 		to.Version,
 		apmErrorLogs{
@@ -53,16 +48,16 @@ func TestUpgrade_8_18_to_8_19_Snapshot(t *testing.T) {
 	}
 }
 
-func TestUpgrade_8_18_to_8_19_BC(t *testing.T) {
+func TestUpgrade_8_17_to_8_18_BC(t *testing.T) {
 	t.Parallel()
-	from := getLatestVersionOrSkip(t, "8.18")
-	to := getLatestBCOrSkip(t, "8.19")
+	from := getLatestVersionOrSkip(t, "8.17")
+	to := getLatestBCOrSkip(t, "8.18")
 	if !from.CanUpgradeTo(to.Version) {
 		t.Skipf("upgrade from %s to %s is not allowed", from.Version, to.Version)
 		return
 	}
 
-	scenarios := basicUpgradeLazyRolloverILMTestScenarios(
+	scenarios := basicUpgradeILMTestScenarios(
 		from.Version,
 		to.Version,
 		apmErrorLogs{
