+++ functionaltests/9_1_test.go
@@ -26,10 +26,10 @@ import (
 // the apm-data plugin update for 8.19 and 9.1:
 // https://github.com/elastic/elasticsearch/pull/119995.
 
-func TestUpgrade_8_18_to_8_19_Snapshot(t *testing.T) {
+func TestUpgrade_9_0_to_9_1_Snapshot(t *testing.T) {
 	t.Parallel()
-	from := getLatestSnapshot(t, "8.18")
-	to := getLatestSnapshot(t, "8.19")
+	from := getLatestSnapshot(t, "9.0")
+	to := getLatestSnapshot(t, "9.1")
 	if !from.CanUpgradeTo(to.Version) {
 		t.Skipf("upgrade from %s to %s is not allowed", from.Version, to.Version)
 		return
@@ -43,6 +43,8 @@ func TestUpgrade_8_18_to_8_19_Snapshot(t *testing.T) {
 			esReturnedUnknown503,
 			refreshCache503,
 			populateSourcemapFetcher403,
+			refreshCache403,
+			refreshCacheESConfigInvalid,
 		},
 	)
 	for _, scenario := range scenarios {
@@ -53,10 +55,10 @@ func TestUpgrade_8_18_to_8_19_Snapshot(t *testing.T) {
 	}
 }
 
-func TestUpgrade_8_18_to_8_19_BC(t *testing.T) {
+func TestUpgrade_9_0_to_9_1_BC(t *testing.T) {
 	t.Parallel()
-	from := getLatestVersionOrSkip(t, "8.18")
-	to := getLatestBCOrSkip(t, "8.19")
+	from := getLatestVersionOrSkip(t, "9.0")
+	to := getLatestBCOrSkip(t, "9.1")
 	if !from.CanUpgradeTo(to.Version) {
 		t.Skipf("upgrade from %s to %s is not allowed", from.Version, to.Version)
 		return
@@ -70,6 +72,66 @@ func TestUpgrade_8_18_to_8_19_BC(t *testing.T) {
 			esReturnedUnknown503,
 			refreshCache503,
 			populateSourcemapFetcher403,
+			refreshCache403,
+			refreshCacheESConfigInvalid,
+		},
+	)
+	for _, scenario := range scenarios {
+		t.Run(scenario.Name, func(t *testing.T) {
+			t.Parallel()
+			scenario.Runner.Run(t)
+		})
+	}
+}
+
+func TestUpgrade_8_19_to_9_1_Snapshot(t *testing.T) {
+	t.Parallel()
+	from := getLatestSnapshot(t, "8.19")
+	to := getLatestSnapshot(t, "9.1")
+	if !from.CanUpgradeTo(to.Version) {
+		t.Skipf("upgrade from %s to %s is not allowed", from.Version, to.Version)
+		return
+	}
+
+	scenarios := basicUpgradeILMTestScenarios(
+		from.Version,
+		to.Version,
+		apmErrorLogs{
+			tlsHandshakeError,
+			esReturnedUnknown503,
+			refreshCache503,
+			populateSourcemapFetcher403,
+			refreshCache403,
+			refreshCacheESConfigInvalid,
+		},
+	)
+	for _, scenario := range scenarios {
+		t.Run(scenario.Name, func(t *testing.T) {
+			t.Parallel()
+			scenario.Runner.Run(t)
+		})
+	}
+}
+
+func TestUpgrade_8_19_to_9_1_BC(t *testing.T) {
+	t.Parallel()
+	from := getLatestVersionOrSkip(t, "8.19")
+	to := getLatestBCOrSkip(t, "9.1")
+	if !from.CanUpgradeTo(to.Version) {
+		t.Skipf("upgrade from %s to %s is not allowed", from.Version, to.Version)
+		return
+	}
+
+	scenarios := basicUpgradeILMTestScenarios(
+		from.Version,
+		to.Version,
+		apmErrorLogs{
+			tlsHandshakeError,
+			esReturnedUnknown503,
+			refreshCache503,
+			populateSourcemapFetcher403,
+			refreshCache403,
+			refreshCacheESConfigInvalid,
 		},
 	)
 	for _, scenario := range scenarios {
